Admin system - PLEASE READ
I've established a system for the admin pages on another branch
(Login-Testing). If a user is logged in in an 'administrator' role, they
will see a link at the top of the page that will take them to the
appropriate admin page (I will send this code to everyone later).
If user is not logged in, or not administrator, they cannot see the
link. If they request an admin page from the url bar, they will be
redirected to the login page.
I'll keep this on the branch for now so it won't be a pain or confusing
while we are developing.